after

+++ api/views.py
@@ -2947,7 +2947,7 @@ def verify_and_do(ref, lieu_depart, lieu_retour, date_depart, heure_depart, date
                         
                         elif credit == "yes":
                             print("Solde (before): ",reservation_obj.client.solde)
-                            reservation_obj.client.solde = reservation_obj.client.solde + Decimal(credit_amount) if reservation_obj.client.solde is not None else Decimal(credit_amount)
+                            reservation_obj.client.solde = reservation_obj.client.solde + float(credit_amount) if reservation_obj.client.solde is not None else float(credit_amount)
                             reservation_obj.save()
                             print("Solde (after): ",reservation_obj.client.solde)
                             sujet_credit = f"SAFAR EL AMIR - Retour anticipé confirmé pour la reservation {reservation_obj.name}"
